commited the poc implementation for replacing addressable asset(prefab) streaming to scene streaming
This is the dirty first implementation as a proposal for addressable scene streaming. There are many hard-to-solve problems with prefabs, for example if there is a prefab instance in a scene having instanced GameObjects or other prefab as its children(when building HLOD data, all these child instances will be lost and only the outmost parent will be left). So, in order to circumvent these issues I propose using scene streaming instead of prefabs. By doing although small amount of memory waste occurs beacuse of cloneing shared parent transforms, most of prefab related issues can be avoided and meanwhile because this way tests each LODGroupds and MeshRenderers(not the prefab) better efficiency is possibly expected while distributing each input GameObject into the matching space node.